+++ check_copyright.sh
@@ -2,7 +2,7 @@
 
 set -eu
 
-die() { echo '$1'; exit 1; }
+die() { echo "$1"; exit 1; }
 
 while read filename; do \
   grep -q "Copyright (c) Facebook" "$filename" ||