+++ usort/tests/__init__.py
@@ -2,3 +2,17 @@
 #
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
+
+from .cli import CliTest
+from .config import ConfigTest
+from .functional import BasicOrderingTest, UsortStringFunctionalTest
+from .sort_key import IsSortableTest, SortableImportTest
+
+__all__ = [
+    "CliTest",
+    "ConfigTest",
+    "BasicOrderingTest",
+    "UsortStringFunctionalTest",
+    "IsSortableTest",
+    "SortableImportTest",
+]