+++ usort/util.py
@@ -4,20 +4,47 @@
 # LICENSE file in the root directory of this source tree.
 
 import os
+from contextlib import contextmanager
 from fnmatch import fnmatch
 from pathlib import Path
-from typing import Generator, Optional
+from time import monotonic
+from typing import Callable, Generator, Iterable, List, Optional, Tuple
 
 import libcst as cst
 
+TIMINGS: List[Tuple[str, float]] = []
 
-def walk(path: Path, glob: str) -> Generator[Path, None, None]:
+
+@contextmanager
+def timed(msg: str) -> Generator[None, None, None]:
+    """
+    Records the monotonic duration of the contained context, with a given description.
+
+    Timings are stored for later use/printing with `print_timings()`.
+    """
+    before = monotonic()
+    yield
+    after = monotonic()
+    TIMINGS.append((msg, after - before))
+
+
+def print_timings(fn: Callable[[str], None] = print) -> None:
+    """
+    Print all stored timing values in microseconds.
+    """
+    for msg, duration in TIMINGS:
+        fn(f"{msg + ':':50} {int(duration*1000000):7} µs")
+
+
+def walk(path: Path, glob: str) -> Iterable[Path]:
+    paths: List[Path] = []
     for root, dirs, files in os.walk(path):
         dirs[:] = [d for d in dirs if not d.startswith(".")]
         root_path = Path(root)
         for f in files:
             if fnmatch(f, glob):
-                yield root_path / f
+                paths.append(root_path / f)
+    return paths
 
 
 def try_parse(path: Path, data: Optional[bytes] = None,) -> cst.Module: