+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## next
+
+* Added support for formatting stdin with `usort format -`
+* Moved diff/check functionality into separate `diff` and `check` commands
+* Replaced `usort format --show-time` with `usort --benchmark` framework
+
 ## 0.5.0a2
 
 * Automatic finding of first-party dirs